requirements.txt
# Natural Language Processing
google.generativeai
langchain-google-genai
python-dotenv
langchain
chromadb
openai

# for visulization
requests
pillow
ipython

# for voice recognition
SpeechRecognition
faster_whisper
pyttsx3
pyaudio

# to chat with our data
docx2txt
unstructured
unstructured[pdf]

#for api
flask